Refactor: consolidate providers
