+++ src/providers/index.ts
@@ -1,0 +1,6 @@
+// ---------
+// PROVIDERS
+// ---------
+
+export * from './highlight';
+export * from './symbols';
